requirements.txt
streamlit
PyPDF2
sqlite3
zipfile
io
re
transformers
pdf2image
base64

